Fixed Issue #125: setuptools leaves a ~/.subversion dir laying around                   after it completes This is the best that can probably be done.  Temporary directories are used for get the svn binary version and to get the initial directory info (for determining if one is, in fact, in a svn working directory)
ALSO: The check for SVN was not right.  Decided on files to only check for
.svn/entries because missing properties just leaves out the external refs.
And then incorporated a test on the code to make sure that svn info completed.

Test passed on CPytonh 2.x and 2.3 on windows, travis-ci, and travis-ci/mac.
Note: There seems to be an issue with pypy and a test.