Preserve order of egg_info section in setup.cfg
egg_info is the dictionary with information that is injected
into setup.cfg. edit_config uses RawConfigParser which uses
collections.OrderedDict for all the data. When we use a
simple dict(), when we loop through items in edit_config, we
see random behavior as a result the fields
tag_svn_revision/tag_date/tag_build are added to the setup.cfg
randomly. So if we sort the items by key when we traverse items
we will get deterministic output as RawConfigParser uses
OrderedDict internally by default.