Prompt for password on upload.
The upload command wasn't prompting for a password. If there was no
password specified in ~/.pypirc, the upload command,
    python setup.py sdist upload
would fail and raise a TypeError. The workaround for this was to use
    python setup.py sdist register upload
since the register command does prompt for a password.

This commit ensures that the upload command prompts for a password
if not given one by ~/.pypirc or the register command.