Ensure shebang lines are correctly quoted if sys.executable contains spaces.
Fixes issue #398. This only special-cases sys.executable; if the --executable
parameter is used, paths with spaces have to be quoted there explicitly. While
this change also applies to Unix platforms, if sys.executable contains spaces
on Unix, any shebang lines created with it aren't going to work either way,
whether they are quoted or not.