Support Python 3.3.0 for extension suffix
In setuptools 27.0+ extension building fails on Python 3.3.0 because the extension suffix is obtained using `SO` and not `EXT_SUFFIX` (which is used in 3.3.1 and above). See: https://hg.python.org/cpython/file/v3.3.0/Lib/distutils/command/build_ext.py#l673

This patch tries `EXT_SUFFIX` and falls back to `SO` if `EXT_SUFFIX` is falsey.