Meaningful message if corrupted setuptools.
Fixes https://github.com/pypa/setuptools/issues/563.

Is not clear in ez_setup.py that it always uses a setuptools archive if
already downloaded: although useful (and even used as a feature in
buildout-bootstrap.py), it's hard for beginners that may have a
corrupted ez_setup.py.

So we just improve the error message, even showing the full path of the
probably "corrupted" zip.