Attempt to fix issue #866 by iterating over code with `dis.Bytecode` instead of the internal `_iter_code`.
The `dis` module was already used in `_iter_code` so I figured it was safe to use `Bytecode` from it. Not sure how this assumption holds up across all supported Python releases. I can only assume `Bytecode` wasn't there before when `_iter_code` was originally written?

Note that `_iter_code` doesn't appear to be called anywhere in light of this change so I removed it.

I should also note that `get_module_constant` has never worked with `setuptools.__version__` (returns -1) because it's not a string literal; it gets that attribute from another module.  But this change does work in cases where a string literal is requested.

https://github.com/pypa/setuptools/issues/866