egg_info: allow specification of setup requirements
This change allows installers such as pip to query
the build-system for setup requirements without having
setuptools first install them. This is not a bulletproof
change that works in every single case, but it is an
improvement because it will eventually allow setup_requires
to be satisfied with wheels, which is the standard
binary format. It also does not introduce a circular
dependency on pip, which could turn out to be a nightmare.