EntryPoint: avoid costly pyparsing for dummy requirement string
In the case where an EntryPoint is created with no extra, a Requirement
object is instanciated with a dummy requirement string 'x[]', which incurs
a useless significant cost due to the use of pyparsing to parse this string.

This change skips the instanciation of a Requirements objects in the
case where extra is empty.

See github issue #1132 for more details, including information on the
performance improvement brought by this change.