Support PEP 345 Project-URL metadata list
A Project-URL list is supported in v1.2 PKG-INFO metadata with
pkginfo 0.6 or later, and is used primarily by PyPI to display
helpful hyperlinks in a generic manner. Add support here to be able
to pass it through setup.py. See the corresponding section of the
Core Metadata Specifications from the Python Packaging User Guide
for details:

https://packaging.python.org/specifications/core-metadata/#project-url-multiple-use