[distutils] patch Python path serach location
Fix the search path to that setuptools
always loads it's own distutils on new
Python versions AND ignore the
vendored distutils on older Pythons.

It's a bit trickier than you might think
and there are some nasty corner cases
that this code needs to handle. This
should cover 99% of cases though and
the rest will probably need a bug report.