Document project_urls setup parameter
Add an entry to the Setuptools usage documentation for the
project_urls dict, and include it in the "advanced" project example.

While at it, adjust the dogfooding use in Setuptools' own setup.py
to use HTTPS for the Documentation URL and drop redundant entries
for Bug Tracker and Source Code.

Also remove a no-op attribute reassignment from egg_info.py.