Read wheel metadata from extracted content instead of from zip
package_name is read from the filename, which can be
case-insensitive (on Windows). zipfile is case-sensitive, so
trying to read the dist-info from it can fail if the filename
case does not match the project name case. Extracting the
files first allows us to read dist-info from the filesystem, which
is also case insensitive on Windows.
Fixes issue #1350