Solve race during creation of egg cache
Upstreaming a private fix that we have applied to python-setuptools 0.9.8 that  solves a race during the creation of an egg cache. We use setuptools 0.9.8 with this fix internally on some of our projects. This fix has been in production for May 2017 and is working great.

The challenge we ran into is that setuptools, by itself, works great, both on development laptops/desktops, our fleet of development VMs, and so on. When put in a high volume CICD pipeline, even the latest version of python-setuptools falls on its face immediately (39.2.0, we haven't tried 40.x since it just came out a few hours ago). 

We suspect that given the build machines have 32 cores, and we're using make -j32, the concurrency is biting us here. We've added this small patch to work around it, and as mentioned, its been great. 

In an effort to get off that ancient version of setuptools, we're submitting this upstream, with the hopes that we can get this mainlined, and can then stay on vanilla "fresh off the branch" versions going forward.