egg_info: Touch 'egg-info' directory
'tox' determines whether a package should be rebuilt by comparing the
timestamp of the package's 'egg-info' directory and its 'setup.py' or
'setup.cfg' files [1][2]. Unfortunately this checks the 'egg-info'
directory itself, which is not updated, unlike the contents of that
directory. This means that 'tox' will always rebuild the package once
one of the two setup files has been updated.

While this is clearly a bug in 'tox' that should be fixed separately,
there is merit in using this as a heuristic so enable it.

[1] https://github.com/tox-dev/tox/blob/3.1.0/src/tox/venv.py#L253-L257
[2] https://github.com/tox-dev/tox/blob/3.1.0/src/tox/venv.py#L221-L244

Signed-off-by: Stephen Finucane <stephen@that.guru>