Make install directory if it doesn't exist yet
Problem: In Fedora, `sudo python3 setup.py install` installs to
/usr/local/lib/pythonX.Y/site-packages. However that directory is
not always there as it is not installed together with the python3
package (Fedora packages should not install stuff to /usr/local).

When the directory does not exist an installation attempt fails with:

  [Errno 2] No such file or directory:
  '/usr/local/lib/python3.7/site-packages/test-easy-install-11875.write-test'

Solution: Make the directory if it doesn't exists yet. If user has
no permissions, the above error is preserved, yet with root (sudo)
it works now.