tox: fix configuration so newer tox versions are supported
Starting with tox 3.2.0, the commands for installing/listing
dependencies changed from pip to python -m pip, which is problematic
as this will leave bits of the previous (from the virtualenv creation)
older version of setuptools installed in the site-packages directory
because from pip's point of view, the active setuptools version is the
one in the current directory (which cannot be uninstalled).

Fix #1453.