Update distribution guide and archive ez_setup doc
This adds a simple introductory guide to the modern way to distribute
`setuptools`-based projects with a link to the Python packaging
tutorial.

Rather than delete the old `ez_setup` information, it has been moved to
a separate archive for people who want to use it as a historical
reference.