Remove erroneous environ commands
`environ.update()` mutates a list in place and returns `None`, so all of
these blocks with `os.environ.copy().update(...)` are creating an
ephemeral copy of the environment dictionary, updating it, and then it's
deleted since nothing holds a reference to it.

Since these were no-ops anyway, I've removed them.