Add setuptools to the build system requirements
Without this, PEP 517 frontends will not be able to find the
setuptools.build_meta backend, which exists only in the source tree.

This presents an obvious bootstrapping problem, in that building
setuptools requires installing setuptools, but at the moment PEP 517 is
read as only supporting build backends that are installable from wheels,
and so we can use the most recent wheel to build this one.