Drop build-backend from pyproject.toml
Because we do not include setuptools in our build requirements and there
is no mechanism for a PEP 517 backend to bootstrap itself, setuptools
cannot use setuptools.build_meta as its backend, and cannot use PEP 517
to build itself.

Additionally, if you specify build-backend, it is impossible to disable
PEP 517 builds with pip's --no-use-pep517 flag.