Preserve encoding with edit_config
PR #1180 and 24be5abd4cb changed the way we read setup.cfg, but when
editing setup.cfg we might still lose the encoding hint on the file,
which can cause failures in subsequent steps.
This change detectes the encoding like in `_parse_config_files` and
writes an encoding hint header if the file encoding was detected.
Fixes https://github.com/jaraco/configparser/issues/37.