Add "python -m setuptools" invocation method
This fully replaces the use of "python setup.py <cmd>" with the
equivalent "python -m setuptools <cmd>". In the process, we guarantee
that a setup.py which imported only distutils will still get the
enhanced setuptools treatment previously provided only by pip install.

This opens the way for future improvements to setuptools, such as
validating setup_requires (if listed in setup.cfg) before invoking the
setup() function and running code.

It is also the only way to invoke setuptools to build and install
packages which fully moved to setup.cfg and no longer provide a stub
setup.py with the content:

```
from setuptool import setup
setup()
```