doc: Update index and follow up on #2093
Build meta was not included originally, emitting a warning from Sphinx
as a result.

In addition, increase the max-depth to 3. As it stands the current table
of contents is overwhelmingly populated from the changelog.

By increasing the amount Sphinx recurses into each individual page, the
content of the site is displayed much more prominently.