Fix repeated command line options handling in distutils.FancyGetopt
Documentation and comments in the code suggest that it shuld be
possible to specify command line options that can be repeated.
However, trying to use this facility for options that take arguments
results in only the last specified argument to be retained. It works
as expected for boolean options. Proably this bug went so far
unnoticed because this functionality is used in setuptools or
distutils only for the '--verbose' option which does not take
arguments.

Fixes #2482