bdist_rpm: add option to force version
The RPM versioning guidelines are slightly at odds with PEP440.
For example, RPM considers 1.1a1 to be a higher version than 1.1
where as PEP440 shows 1.1a1 to be an alpha and 1.1 to be the
official release.
Permitting specific override of the RPM version will allow a
maintainer to explicitly correct the version to fit their needs.