Add default values for `packages` and `py_modules`
The need for explicitly setting config options for `packages`,
`py_modules` and `package_dir` is usually mentioned as a pitfall of
setuptools and can be confusing some times.

The changes introduced here explore the conventions of the Python
ecosystem (namely ``flat-layout`` and ``src-layout`` for packages) to
provide default values for these configurations.

Notice that the default values will only be applied if the `py_modules`
or `packages` options are not given. This is intentional, since the
implementation is supposed to be backwards compatible.

References:

https://blog.ionelmc.ro/2015/02/24/the-problem-with-packaging-in-python/
https://blog.ionelmc.ro/2014/06/25/python-packaging-pitfalls/