Add `ini2toml` as a vendored dependency
Some `ini2toml` modules are removed, since they are not necessary for
setuptools use case and might require extra dependencies (e.g. CLI
usage, writing TOML files, format preserving).

--

The automatic conversion of `setup.cfg` into `pyproject.toml` as the
*one true way* of doing declarative builds was first suggested in #1688.

--

There are advantages and disadvantages in using an external tool such
as `ini2toml` for the automatic conversion.

The main advantage is that users can use the exact same tool for
converting their old packages and it will work in the same way
setuptools works. This means that on top of doing an automatic
conversion, the users are also offered an alternative to explicitly
update their package configuration.

The main disadvantage is that `ini2toml` works in a way that is a bit
more complex than simply parsing the `setup.cfg` file and dumping it as
`pyproject.toml` (we can think about `ini2toml` as if it was
transforming an "AST-equivalent" representation of the INI and TOML
formats -- this is necessary for other use cases covered by the library).
In order to minimise this complexity, some parts of the
package are stripped out during the vendoring process.

Also note that currently `ini2toml` ships a "built-in" plugin for
setuptools. In the future, if there is interest, there is also the
possibility of moving this plugin directly under setuptools repository.