Use run_setup to obtain a dist object and apply the new configs
The `run_setup` function from `distutils.core` allows retrieving the
distribution object before the configuration is read.
Then using the new functions from the `setuptools.config` package
we can apply the TOML-style configuration (either from a
`pyproject.toml` file or by an automatically converted `setup.cfg`).

To ensure backward compatibility, in the case of discrepancies in the
conversion of the `setup.cfg` files, the implementation falls back to
the legacy configuration procedure.