Make include_package_data=True for `pyproject.toml` configs
There seems to be an opinion in the community that
`include_package_data=True` would be a better default and a quality of
life improvement.

This change relies on the fact that `ini2toml` will automatically
backfill `include_package_data=False` for converted `setup.cfg` files
to implement this change in a backward compatible way.