cygwinccompiler: Split CC env var before passing to subprocess
4113bc31a8e62 added support for clang by respecting the CC env variable.
The content of the env var gets passed to check_output() as the compiler
executable. But CC is not just a path to an executable but a command line,
such as "ccache gcc" which makes checkout_output() fail in those cases because
it will try to look for "ccache gcc" in PATH instead of "ccache".

To fix the issue use shlex to parse the command line before passing it to
check_output().