msvc9compiler: Don't raise DistutilsPlatformError on import
It gets raised when get_build_version() returns a too low version
or when it fails to detect a MSVC version, which for exaple is the
case when CPython is built with gcc/clang.

In theory this module isn't needed on non-MSVC platforms, but setuptools
imports it anyway when monkey patching, which in turn makes setuptools fail.

To work around this only check the version when initialising MSVCCompiler().

This also mirrors the behaviour of the newer MSVCCompiler() (in _msvccompiler)
which also raises DistutilsPlatformError() only on initialisation in case
it can't find the right MSVC.