sysconfig: use get_config_h_filename() from the stdlib
distutils.sysconfig provides various functions already present in
the stdlib sysconfig module.

Reusing the stdlib versions has the advantage that it is in control of
the Python distributor/packager and distutils doesn't have to deal with
platform details.

This tries to start the transition using the simple get_config_h_filename()
for starters by forwarding calls to the stdlib in case Python is
considered to be installed.