tests: fix tests on Ubuntu 22.04
I added a test for sysconfig.parse_config_h() in 9d0b8cda407
which assumed pyconfig.h always has some macros defined and checked
that the result was non-empty.

On Ubuntu 22.04 the pyconfig.h is just a shim, including various platform specific
configs and not defining anything directly.

This changes the test to not assume anything about the output of parse_config_h() instead.