Remove `numpy` sdists from integration tests
Numpy now uses a version cap for setuptools in their `pyproject.toml`,
which defies a bit the value of including it in the integration tests.

As revealed in a conversation with `numpy` maintainer, it seems that
they not plan to remove this cap (maybe only update it from time to
time, only if necessary). Moreover they are also studying other
build backends.

So I think that now the best thing to do would be stop trying to build
numpy sdists with the current version of setuptools, otherwise we risk
to break this test.