Use meta path finder instead for pytest plugins
When pytest runs on PyPy, the previous explored approach
(i.e. renaming setuptools/tests/{fixtures => conftest}.py)
fails.

In this commit we explore the usage of a meta path finder to make sure
pytest can load in-tree plugins (and test helpers) even when setuptools
is installed from a distribution files (wheel or sdist).