Fix codecov config
The previous `.codecov.yml` file had an invalid configuration:

```bash
$ curl --data-binary @.codecov.yml https://codecov.io/validate
Error at ['coverage', 'status', 'project', 'threshold']:
must be of ['dict', 'boolean'] type
```

According to the
[documentation](https://docs.codecov.com/docs/commit-status#section-project-status),
`coverage.status.project` does not seem to accept configuration
parameters directly. Instead it requires a mapping between project
names (or `"default"` when all the projects are target) and the
configuration parameters.

This change fixes that by introducing a new nesting level with the
`default` key between `project` and `default`. This passes the validation:

```bash
% curl --data-binary @.codecov.yml https://codecov.io/validate
Valid!

{
  "comment": false,
  "coverage": {
    "status": {
      "project": {
        "default": {
          "threshold": 0.5
        }
      }
    }
  }
}
```