Reduce GitHub Actions timeout from 6h to 75min
From times to times, tests (specially on PyPy) seem to get
stuck and be cancelled by GitHub after the default timeout (6h).

Before these tests are cancelled, however they count as an active job
and limit the amount of concurrent jobs/workflows PyPA's organization
may have running.

Currently, the slowest job running for the `main` workflow seem to be the
tests for PyPy+Windows, which take around 55 min. Therefore, chances are
that, if a test is taking much more than 1h to run, it got stuck.

We can be pragmatic and reduce the timeout for the `main`
workflow and free the resources quickly.