Disable coverage on PyPy
Currently, CI jobs running on PyPy are (painfully) slow.
This seems to be a well know side effect of enabling coverage.

The change proposed here is to disable coverage on PyPy for both Windows
and MacOS (which seem to be very slow).

Coverage for PyPy on Ubuntu is preserved because, while
being slower than CPython, it is still OK-ish.
Hopefully tests running on PyPy/Ubuntu will include most of
PyPy's specific corner cases.