Don't warn on false positive for author/maintainer's email
While I was working to support pyproject.toml metadata in setuptools,
I received as a feedback from the community[^1] that setuptools
warns the following message when `author_email` and `maintainer_email`
are given in the form of `Person Name <email@address>`:

> warning: check: missing meta-data: either (author and author_email)
> or (maintainer and maintainer_email) should be supplied

This can be seen as a false positive, because indeed both author's name
and email are provided.

This warning seems to happen because distutils define the `check`
command as a subcommand for `sdist`.

This change aims to remove this false positive result from the checks.

[^1]: https://discuss.python.org/t/help-testing-experimental-features-in-setuptools/13821/18