Attempt to clarify which url is missing for pyproject-metadata builds
When the user does not specify `Homepage` (or any variant such as
`home-page`), distutils will warn the following message:

    warning: check: missing required meta-data: url

This message is fine for `setup.cfg` builds because the field there is
called `url`, but it does not work well for builds using pyproject.toml
metadata.

The change implemented here will add some other logging information that
try to point the user in the correct direction for solving this issue.

This problem was first identified in:
https://discuss.python.org/t/help-testing-experimental-features-in-setuptools/13821