Restructure config.pyproject to consider "pre-set" dynamic values
Issues 3195 and 3204 surface the fact that setuptools may need to allow
dynamic values to be computed by the users in the ``setup.py`` file
(e.g. if they need to dynamically decide dependencies based on the host
machine in a way that is not supported by environment markers, such as
GPU presence).

The current implementation somehow already allows that by layering the
configs `setup.py` > `setup.cfg` > `pyproject.toml`. However this is
done without having in mind the limitations about `dynamic` imposed by
PEP 621.

The change implemented here tries to fix this problem.