Fix duplicated version tags in egg_info
Previously egg_info was adding duplicated tags to the version string.
This was happening because of the version normalization.
When the version normalization was applied to the string the tag was
modified, then later egg_info could no longer recognize it before
applying.

The fix for this problem was to normalize the tag string before
applying.