Clarify that `pyproject.toml` config will not be removed
The current documentation regarding `pyproject.toml` claims that
support for using it to specify package metadata and build
configuration options might be "completely removed" in a future
release.

However, nowadays it is clear
(https://github.com/pypa/setuptools/issues/1688#issuecomment-1079706929
and #3214) that `pyproject.toml` is not going anywhere, is here
to stay, and is in fact expected to be the preferred format
in the future.

This makes an incremental change toward that future by clarifying
that `pyproject.toml` support will not be removed in a future
release of setuptools.