Added two `mypkg` directories in the example directory trees
Since the previous example on manually specifying `packages` and
`package_dir` alludes to two packages, `mypkg1` and `mypkg2`, IMHO it
would be more consistent to continue this and show two packages in the
directory trees for the src-layout and flat-layout, instead of showing a
single package `mypkg`. This would also make it more clear that multiple
packages are allowed.