Section on Entry Points Syntax
Matter is courtesy of @abravalheri.

Have removed the `.ini` syntax highlighting of the code blocks so that
the syntax does not appear specific to `setup.cfg` but more
general-purpose.

Have also moved this section down to the end of the document so that
users can refer to it after having a basic understanding of how each of
console scripts, GUI scripts, and entry points work.